+++ web/modules/custom/hoeringsportal_base_fixtures/src/Fixture/PublicMeetingFixture.php
@@ -369,10 +369,10 @@ final class PublicMeetingFixture extends AbstractFixture implements DependentFix
             'email' => $spec['email'] ?? 'test@example.com',
             'positions' => array_map(
               static fn(int $index) => [
-                'subevent' => $subEventInfo['pretix_subevent_id'],
+                'subevent' => $subEventInfo['pretix_subevent_id'] ?? [],
                 'attendee_name' => sprintf('Attendee %d', $index),
                 'attendee_email' => sprintf('test%03d@example.com', $index),
-                'item' => reset($subEventInfo['data']['subevent']['item_price_overrides'])['item'],
+                'item' => isset($subEventInfo['data']['subevent']['item_price_overrides']) ? reset($subEventInfo['data']['subevent']['item_price_overrides'])['item'] : FALSE,
               ],
               range(1, $line['quantity'] ?? 1),
             ),