+++ web/themes/custom/hoeringsportal/assets/js/hoeringsportal.js
@@ -17,6 +17,9 @@ require("slick-carousel");
 require("./slick-slider-config.js");
 require("./filter-collapse.js");
 require("./icons.js");
+require("./modify-dialogue-form.js");
+require("./modify-dialogue-proposal-comments.js");
+require("./animated-svg.js");
 
 // Enable popovers.
 $(function () {