+++ CHANGELOG.md
@@ -6,6 +6,19 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+* [PR-553](https://github.com/itk-dev/deltag.aarhus.dk/pull/553)
+  * Add animated svg icons comment, reply, thumbs-up
+
+* [PR-552](https://github.com/itk-dev/deltag.aarhus.dk/pull/552)
+  * Add styling for dialog proposal categories
+
+* [PR-551](https://github.com/itk-dev/deltag.aarhus.dk/pull/551)
+  * Remove old video field and disabled Azure video module
+  * Enable Comments module
+  * Add Support comment flag
+  * Add Honeypot module
+  * Add ITK Video module and Video paragraph for Video Tool
+
 ## [4.13.1] - 2025-09-15
 
 * [PR-548](https://github.com/itk-dev/deltag.aarhus.dk/pull/548)